+++ bin/get-hosts.py
@@ -40,9 +40,9 @@ if __name__ == '__main__':
 
     parser.add_argument(
         'ids', nargs='*',
-        help=('The accession number(s) or taxonomy id(s) to print host '
-              'information for. If not given, accession numbers are read '
-              'from standard input, one per line.'))
+        help=('The ids (accession numbers, names, or taxonomy ids) to print '
+              'host information for. If not given, ids are read from '
+              'standard input, one per line.'))
 
     parser.add_argument(
         '--database', required=True,